+++ packages/interfaces/src/misc.ts
@@ -1,0 +1,11 @@
+import type { IDecodedMessage } from "./message.js";
+
+export interface IAsyncIterator<T extends IDecodedMessage> {
+  iterator: AsyncIterator<T>;
+  stop: Unsubscribe;
+}
+
+export type Unsubscribe = () => void | Promise<void>;
+
+export type PubSubTopic = string;
+export type ContentTopic = string;