+++ packages/utils/src/common/index.ts
@@ -1,3 +1,4 @@
 export * from "./is_defined.js";
 export * from "./random_subset.js";
 export * from "./group_by.js";
+export * from "./to_async_iterator.js";